Expanding TclOO method names in tkcon.
Doing more TclOO stuff, I wanted to be able to expand method names interactively in the same manner that proc names are expanded. Turned out the tkcon already had that for nx, xotcl, nsf so I patched in support for TclOO.

The patch also fixes some limitations in the original code. In particular, it also expands method names when the cursor is not at the end of the line, or within a multi command line like first command ; obj methodcall or when the object is accessed via a variable like $obj methodcall.

See http://wiki.tcl.tk/41346
and http://sourceforge.net/p/tkcon/feature-requests/9/