Update tkcon.tcl and README.txt
Bump version to V2.7.10.  I was unable to find a "most-recent-minor-version", so I pushed to 2.7.10.

Added handling of ctrl-plus and ctrl-minus for resizing console fonts, similar to the pattern used in most web-browsers.
Added handling of control-plus and control-minus keypresses to increase and decrease font sizes in the consoles.

On Win10 platforms the menus are handled by the native system so changing font sizes is non-trivial without a significant rewrite of the TkCon code.from